Refactor Redux Store Structure and Move Features to Separate Directories
- Removed hooks.ts and moved its content to a new hooks directory.
- Reorganized store structure by creating feature directories for layout, pokemon, and styleSettings.
- Deleted old slice files and created new ones in their respective feature directories.
- Updated index files to export reducers, actions, and selectors from the new feature directories.
- Implemented new selectors and thunks for the pokemon feature.
- Added type definitions for layout and style settings in separate model files.
- Ensured local storage persistence for layout and style settings.
