+++ src/hooks/useViewMode.ts
@@ -1,5 +1,6 @@
-import { useAppSelector, useAppDispatch } from '../store/hooks'
-import { setViewMode as setViewModeAction, type ViewMode } from '../store/styleSettingsSlice'
+import { useAppSelector, useAppDispatch } from '@/store/hooks'
+import { setViewMode as setViewModeAction } from '@/store/features/styleSettings'
+import type { ViewMode } from '@/models/store/StylesSetting'
 
 export function useViewMode() {
 	const viewMode = useAppSelector((state) => state.styleSettings.viewMode)
