+++ src/models/store/Layout.ts
@@ -1,0 +1,5 @@
+export type LayoutType = 'empty' | 'header'
+
+export interface LayoutState {
+	layoutType: LayoutType
+}
