+++ src/store/features/layout/index.ts
@@ -1,0 +1,14 @@
+// ===================================
+// REDUCER
+// ===================================
+export { default as layoutReducer } from './layoutSlice'
+
+// ===================================
+// ACTIONS
+// ===================================
+export { setLayoutType } from './layoutSlice'
+
+// ===================================
+// SELECTORS
+// ===================================
+export { selectLayoutType, selectLayoutState } from './selectors'
