+++ src/store/features/layout/thunks.ts
@@ -1,0 +1,2 @@
+// Layout feature no necesita thunks ya que solo maneja estado local
+// Si en el futuro se necesitan operaciones asíncronas, se pueden agregar aquí
