+++ src/store/features/styleSettings/thunks.ts
@@ -1,0 +1,2 @@
+// StyleSettings feature no necesita thunks ya que solo maneja configuraciones locales
+// Si en el futuro se necesitan operaciones asíncronas (como guardar en servidor), se pueden agregar aquí
