+++ src/store/hooks/hooks.ts
@@ -1,5 +1,5 @@
 import { useDispatch, useSelector, type TypedUseSelectorHook } from 'react-redux'
-import type { RootState, AppDispatch } from './index'
+import type { RootState, AppDispatch } from '../index'
 
 // Use throughout your app instead of plain `useDispatch` and `useSelector`
 export const useAppDispatch = () => useDispatch<AppDispatch>()
